+++ R/n_obs.R
@@ -74,8 +74,7 @@ n_obs.glm <- function(x, ...) {
         .nobs <- sum(resp_data)
       }
     } else if (!is.data.frame(resp_data) && !.is.int(resp_data)) {
-      w <- get_weights(x)
-      .nobs <- sum(resp_data * w)
+      .nobs <- sum(get_weights(x))
     }
   }
 
@@ -184,6 +183,13 @@ n_obs.bayesx <- function(x, ...) {
 #' @export
 n_obs.flexsurvreg <- function(x, ...) {
   x$N
+}
+
+
+
+#' @export
+n_obs.SemiParBIV <- function(x, ...) {
+  x$n
 }